requirements.txt
# TrufflePiggie Dependencies
# ==========================

# HTTP requests with retry support
requests>=2.31.0
urllib3>=2.0.0

# Beautiful CLI output
rich>=13.7.0

# Configuration management
PyYAML>=6.0.1

# Type hints (for Python < 3.10)
typing-extensions>=4.9.0

